Added result cards for the adams-bohart and yoon-nelson models
